move resetDehydratedStateTimeStamp to utils.ts
