+++ packages/commerce-sdk-react/src/utils.ts
@@ -7,6 +7,7 @@
 
 import Cookies, {CookieAttributes} from 'js-cookie'
 import {IFRAME_HOST_ALLOW_LIST} from './constant'
+import { DehydratedState } from '@tanstack/react-query'
 
 /** Utility to determine if you are on the browser (client) or not. */
 export const onClient = (): boolean => typeof window !== 'undefined'
@@ -109,5 +110,32 @@ export function detectCookiesAvailable(options?: CookieAttributes) {
         return success
     } catch {
         return false
+    }
+}
+
+/**
+ * Resets the dataUpdatedAt timestamp for all mutations and queries in the given data object.
+ * This is typically used in conjuction with pwa-kit-react-sdk's`beforeHydrate` to ensure that
+ * the cached data is considered fresh on first load.
+ * @param data - The data object containing mutations and queries.
+ * @param timestamp - The timestamp to set the dataUpdatedAt to. If not provided, the current time will be used.
+ * @returns The updated data object with reset timestamps.
+ * @since 4.0.0
+ * @public
+ */
+export const resetDehydratedStateTimeStamp = (data: DehydratedState, timestamp?: Date) => {
+    const time = timestamp || Date.now()
+
+    const updateQueryTimeStamp = (item: any) => ({
+        ...item,
+        state: {
+            ...item.state,
+            dataUpdatedAt: time
+        }
+    })
+
+    return {
+        mutations: data.mutations.map(updateQueryTimeStamp),
+        queries: data.queries.map(updateQueryTimeStamp)
     }
 }
